+++ src/basic/App.tsx
@@ -1,10 +1,12 @@
 import { useState } from "react";
-import { CartItem } from "../types";
+import { CartItem, Coupon, ProductWithUI } from "../types";
 import { Header } from "./components/common/Header";
 import { Notification } from "../types";
 import { NotificationToast } from "./components/common/NotificationToast";
 import CartPage from "./pages/CartPage";
 import AdminPage from "./pages/AdminPage";
+import { initialCoupons, initialProducts } from "./constants";
+import { useAddNotification } from "./hooks/notification/useAddNotification";
 
 export default function App() {
   const [cart, setCart] = useState<CartItem[]>(() => {
@@ -18,10 +20,37 @@ export default function App() {
     }
     return [];
   });
+  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
   const [isAdmin, setIsAdmin] = useState(false);
   const [notifications, setNotifications] = useState<Notification[]>([]);
   const [searchTerm, setSearchTerm] = useState("");
   const [totalItemCount, setTotalItemCount] = useState(0);
+
+  const { addNotification } = useAddNotification(setNotifications);
+
+  const [coupons, setCoupons] = useState<Array<Coupon>>(() => {
+    const saved = localStorage.getItem("coupons");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialCoupons;
+      }
+    }
+    return initialCoupons;
+  });
+
+  const [products, setProducts] = useState<ProductWithUI[]>(() => {
+    const saved = localStorage.getItem("products");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialProducts;
+      }
+    }
+    return initialProducts;
+  });
 
   return (
     <div className="min-h-screen bg-gray-50">
@@ -37,18 +66,32 @@ export default function App() {
         cart={cart}
         totalItemCount={totalItemCount}
       />
-      {isAdmin ? (
-        <AdminPage />
-      ) : (
-        <CartPage
-          cart={cart}
-          isAdmin={isAdmin}
-          setCart={setCart}
-          searchTerm={searchTerm}
-          setTotalItemCount={setTotalItemCount}
-          setNotifications={setNotifications}
-        />
-      )}
+      <main className="max-w-7xl mx-auto px-4 py-8">
+        {isAdmin ? (
+          <AdminPage
+            coupons={coupons}
+            setCoupons={setCoupons}
+            products={products}
+            setProducts={setProducts}
+            selectedCoupon={selectedCoupon}
+            setSelectedCoupon={setSelectedCoupon}
+            addNotification={addNotification}
+            cart={cart}
+          />
+        ) : (
+          <CartPage
+            cart={cart}
+            setCart={setCart}
+            products={products}
+            coupons={coupons}
+            selectedCoupon={selectedCoupon}
+            setSelectedCoupon={setSelectedCoupon}
+            searchTerm={searchTerm}
+            setTotalItemCount={setTotalItemCount}
+            addNotification={addNotification}
+          />
+        )}
+      </main>
     </div>
   );
 }
